// frontend-admin/src/api/generated/model/licenseInfo.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export interface LicenseInfo {
  /** @nullable */
  canonicalLicenseKey?: string | null;
  /** @nullable */
  customerName?: string | null;
  exists?: boolean;
  isExpired?: boolean;
  isRevoked?: boolean;
  isValid?: boolean;
  /** @nullable */
  licenseKey?: string | null;
  /** @nullable */
  licenseKind?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  sourceId?: string | null;
  /** @nullable */
  sourceTable?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tenantId?: string | null;
  /** @nullable */
  tenantSlug?: string | null;
  /** @nullable */
  validUntilUtc?: string | null;
}